- changes IGDB sync jobs from BullMQ to Promise-based approach ...probably has drawbacks...
