+++ src/sync/igdb/igdb-sync.module.ts
@@ -1,29 +1,13 @@
 import { Module } from "@nestjs/common";
 import { IgdbSyncService } from "./igdb-sync.service";
-import { BullModule } from "@nestjs/bullmq";
-import { IgdbSyncProcessor } from "./igdb-sync.processor";
-import { IGDB_SYNC_QUEUE_NAME } from "./igdb-sync.constants";
 import { GameRepositoryModule } from "../../game/game-repository/game-repository.module";
 
 /**
- * This module is responsible for handling the game create/update queue.
- * The game-node-sync-igdb service is responsible for sending requests to
- * this module's controller, with a list of games returned from the IGDB API.
+ * This module is responsible for handling game create/update with data
+ * received from game-node-sync-igdb
  */
 @Module({
-    imports: [
-        BullModule.registerQueue({
-            name: IGDB_SYNC_QUEUE_NAME,
-            defaultJobOptions: {
-                // If this is not used, Redis will take a lot of ram for completed jobs
-                removeOnComplete: true,
-                removeOnFail: true,
-                attempts: 1,
-                backoff: 300,
-            },
-        }),
-        GameRepositoryModule,
-    ],
-    providers: [IgdbSyncService, IgdbSyncProcessor],
+    imports: [GameRepositoryModule],
+    providers: [IgdbSyncService],
 })
 export class IgdbSyncModule {}
